Note bug, add tracing info, bake into module
Testing revealed a bug in the case where we collect more than 53 bits
worth of significance. I have vague understanding of why that happens
but I'll just note the bug and apply a workaround for now.

Add state printouts as tracing info when compiled with DEBUG directive.

Change private things to private etc. to make this more modulelike.